+++ dev-utils-service/src/main/java/com/ttg/devknowledgeplatform/devutils/service/impl/CsvToJsonOperation.java
@@ -19,6 +19,7 @@ import com.ttg.devknowledgeplatform.common.exception.BusinessException;
 import com.ttg.devknowledgeplatform.devutils.exception.DevUtilsErrorCode;
 import com.ttg.devknowledgeplatform.devutils.exception.ParsingExceptionMessages;
 import com.ttg.devknowledgeplatform.devutils.service.DevUtilOperation;
+import com.ttg.devknowledgeplatform.devutils.service.OperationGroup;
 import com.ttg.devknowledgeplatform.devutils.service.impl.support.JsonNodeIo;
 
 import lombok.RequiredArgsConstructor;
@@ -54,6 +55,11 @@ import lombok.RequiredArgsConstructor;
 public class CsvToJsonOperation implements DevUtilOperation {
 
     private final ObjectMapper objectMapper;
+
+    @Override
+    public OperationGroup group() {
+        return OperationGroup.FORMATTERS;
+    }
 
     /**
      * @throws BusinessException wrapping {@link DevUtilsErrorCode#INVALID_CSV} when