+++ dev-utils-service/src/main/java/com/ttg/devknowledgeplatform/devutils/service/impl/JsonToCsvOperation.java
@@ -18,6 +18,7 @@ import com.fasterxml.jackson.dataformat.csv.CsvSchema;
 import com.ttg.devknowledgeplatform.common.exception.BusinessException;
 import com.ttg.devknowledgeplatform.devutils.exception.DevUtilsErrorCode;
 import com.ttg.devknowledgeplatform.devutils.service.DevUtilOperation;
+import com.ttg.devknowledgeplatform.devutils.service.OperationGroup;
 import com.ttg.devknowledgeplatform.devutils.service.impl.support.JsonNodeIo;
 
 import lombok.RequiredArgsConstructor;
@@ -64,6 +65,11 @@ import lombok.RequiredArgsConstructor;
 public class JsonToCsvOperation implements DevUtilOperation {
 
     private final ObjectMapper objectMapper;
+
+    @Override
+    public OperationGroup group() {
+        return OperationGroup.FORMATTERS;
+    }
 
     /**
      * @throws BusinessException wrapping {@link DevUtilsErrorCode#INVALID_JSON} when