+++ dev-utils-service/src/main/java/com/ttg/devknowledgeplatform/devutils/service/impl/PhpToJsonOperation.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.ttg.devknowledgeplatform.common.exception.BusinessException;
 import com.ttg.devknowledgeplatform.devutils.exception.DevUtilsErrorCode;
 import com.ttg.devknowledgeplatform.devutils.service.DevUtilOperation;
+import com.ttg.devknowledgeplatform.devutils.service.OperationGroup;
 import com.ttg.devknowledgeplatform.devutils.service.impl.support.JsonNodeIo;
 import com.ttg.devknowledgeplatform.devutils.service.impl.support.PhpArrayParser;
 
@@ -29,6 +30,11 @@ import lombok.RequiredArgsConstructor;
 public class PhpToJsonOperation implements DevUtilOperation {
 
     private final ObjectMapper objectMapper;
+
+    @Override
+    public OperationGroup group() {
+        return OperationGroup.FORMATTERS;
+    }
 
     /**
      * @throws BusinessException wrapping {@link DevUtilsErrorCode#INVALID_PHP} when {@code input}